+++ docs/conventions.rst
@@ -69,12 +69,12 @@ is based on the reference API that's being implemented.
 
 As an example, the BEA API, implemented by :mod:`finagg.bea.api`, contains
 a singleton :data:`finagg.bea.api.gdp_by_industry` with an attribute
-:data:`finagg.bea.api.gdp_by_industry.name` that describes the BEA API database
+:attr:`~finagg.bea.api.API.name` that describes the BEA API database
 that the singleton refers to. In addition, the singleton has methods
-:data:`finagg.bea.api.gdp_by_industry.get_parameter_list` and
-:data:`finagg.bea.api.gdp_by_industry.get_parameter_values`
+:meth:`~finagg.bea.api.API.get_parameter_list` and
+:meth:`~finagg.bea.api.API.get_parameter_values`
 for getting API parameters and API parameter value options, respectively,
-while :data:`finagg.bea.api.gdp_by_industry.get` is the actual method for
+while :meth:`~finagg.bea.api.GDPByIndustry.get` is the actual method for
 retrieving data from the API is implemented by.
 
 >>> finagg.bea.api.gdp_by_industry.name
@@ -111,11 +111,11 @@ IDs available through other methods. These methods are useful for verifying
 if data has been installed properly or for selecting a subset of data for
 further refinement. Examples of these methods include:
 
-* :meth:`finagg.fred.feat.series.get_id_set` returns installed economic data
+* :meth:`finagg.fred.feat.raw.Series.get_id_set` returns installed economic data
   series IDs
 * :meth:`finagg.sec.api.get_ticker_set` returns all the tickers that have
   at least *some* data available through the SEC EDGAR API
-* :meth:`finagg.sec.feat.quarterly.get_ticker_set` returns all the tickers
+* :meth:`finagg.sec.feat.refined.quarterly_.Quarterly.get_ticker_set` returns all the tickers
   that have quarterly features available
 
 Data organization
@@ -131,9 +131,7 @@ There are only a handful of conventions regarding data organization:
   per feature dataframe column. This makes it so features can be changed without
   breaking the SQL table schemas.
 * Classes within ``feat`` submodules and SQL tables within ``sql`` submodules are
-  named similarly to indicate their relationship. As an example,
-  :data:`finagg.sec.sql.quarterly` corresponds to the SQL table definition for
-  :data:`finagg.sec.feat.quarterly`.
+  named similarly to indicate their relationship.
 * Unaltered data from APIs are typically referred to as "raw" data while
   features are referred to as "refined" data. Refined data SQL tables typically
   have foreign key constraints on raw data SQL tables such that refined rows
