+++ src/finagg/yfinance/sql.py
@@ -15,6 +15,12 @@ prices = sa.Table(
     sa.Column("close", sa.Float, doc="Stock price at market close."),
     sa.Column("volume", sa.Float, doc="Units traded during trading hours."),
 )
+"""SQL table for storing raw data as managed by
+:data:`finagg.yfinance.feat.prices` (an alias for
+:class:`finagg.yfinance.feat.Prices`).
+
+:meta hide-value:
+"""
 
 daily = sa.Table(
     "yfinance.refined.daily",
@@ -30,3 +36,9 @@ daily = sa.Table(
     sa.Column("name", sa.String, primary_key=True, doc="Feature name."),
     sa.Column("value", sa.Float, nullable=False, doc="Feature value."),
 )
+"""SQL table for storing refined data as managed by
+:data:`finagg.yfinance.feat.daily` (an alias for
+:class:`finagg.yfinance.feat.Daily`).
+
+:meta hide-value:
+"""
